ElectricityMarketData

Documentation for ElectricityMarketData.

ElectricityMarketData.ElectricityMarket

TimeZones.ZonedDateTime

ElectricityMarketData._async_download

ElectricityMarketData._async_download

ElectricityMarketData._async_get_raw_data

ElectricityMarketData._download

ElectricityMarketData._get_data

ElectricityMarketData._get_raw_data

ElectricityMarketData.available_time_series

ElectricityMarketData.available_time_series

ElectricityMarketData.available_time_series

ElectricityMarketData.get_acess_key_headers

ElectricityMarketData.get_day_ahead_hourly_lmp

ElectricityMarketData.get_day_ahead_hourly_lmp_raw_data

ElectricityMarketData.get_day_ahead_lmp

ElectricityMarketData.get_day_ahead_lmp

ElectricityMarketData.get_day_ahead_lmp_raw_data

ElectricityMarketData.get_day_ahead_lmp_raw_data

ElectricityMarketData.get_group_df

ElectricityMarketData.get_meta_df

ElectricityMarketData.get_real_time_lmp

ElectricityMarketData.get_real_time_lmp

ElectricityMarketData.get_real_time_lmp_raw_data

ElectricityMarketData.get_real_time_lmp_raw_data

ElectricityMarketData.get_str_dates

ElectricityMarketData.get_timezone

ElectricityMarketData.get_timezone

ElectricityMarketData.get_timezone

ElectricityMarketData.get_url_day_ahead_hourly_lmp

ElectricityMarketData.list_markets

ElectricityMarketData.parse_df_format

ElectricityMarketData.ElectricityMarket — Type

ElectricityMarket

Abstract type representing an electricity market.

TimeZones.ZonedDateTime — Method

ZonedDateTime(date::DateTime, timezone::TimeZone) :: ZonedDateTime

Return a ZonedDateTime object with the given date and timezone from the market.

ElectricityMarketData._async_download — Method

_async_download(url::AbstractString, filename::AbstractString)::Task

Downloads the 'url' and saves it to 'filename' asynchronously. Return the tasks.

ElectricityMarketData._async_download — Method

function _async_download(urls::Vector{T}, filenames::Vector{T}, url_zip::AbstractString, filename_zip::AbstractString)::Vector{Task} where T <: AbstractString

Downloads the 'urls' and saves it to 'filenames' asynchronously. If the file is not available, uses the zipped version. Return the tasks.

ElectricityMarketData._async_get_raw_data — Method

_async_get_raw_data(market::MisoMarket, url_folder::AbstractString, file_base::AbstractString, extension::AbstractString, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir())

Download the 'filebase' raw data for the given market in the given 'urlfolder' and start_date to end_date and save it in folder.

ElectricityMarketData._download — Function

_download(url::AbstractString, filename::AbstractString)::AbstractString

Downloads the return of url if the file filename does not exist, and saves it in filename.

ElectricityMarketData._get_data — Method

_get_data(market::MisoMarket, url_folder::AbstractString, file_base::AbstractString, extension::AbstractString, type::AbstractString, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Real-Time (RT) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData._get_raw_data — Method

_get_raw_data(market::MisoMarket, url_folder::AbstractString, file_base::AbstractString, extension::AbstractString, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Real-Time (RT) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder.

ElectricityMarketData.available_time_series — Method

available_time_series(::MisoMarket) :: Vector{NamedTuple}

Return Vector of NamedTuple of available time series for the given market.

Ex:

[
    (name="RT-load", unit="MW", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_load_data, description="Real-time load data"),
    (name="RT-LMP", unit="MWh", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_lmp, description="Real-time Locational Marginal Price data"),
]

ElectricityMarketData.available_time_series — Method

available_time_series(::PjmMarket) :: Vector{NamedTuple}

Return Vector of NamedTuple of available time series for the given market.

Ex:

[
    (name="RT-load", unit="MW", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_load_data, description="Real-time load data"),
    (name="RT-LMP", unit="MWh", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_lmp, description="Real-time Locational Marginal Price data"),
]

ElectricityMarketData.available_time_series — Method

available_time_series(market::ElectricityMarket) :: Vector{NamedTuple}

Return Vector of NamedTuple of available time series for the given market.

Ex:

[
    (name="RT-load", unit="MW", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_load_data, description="Real-time load data"),
    (name="RT-LMP", unit="MWh", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_lmp, description="Real-time Locational Marginal Price data"),
]

ElectricityMarketData.get_acess_key_headers — Method

get_acess_key_headers()::Dict

Get the access key headers from the PJM API.

ElectricityMarketData.get_day_ahead_hourly_lmp — Method

get_day_ahead_hourly_lmp(market::PjmMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Return a table with data for Day-Ahead Energy Market locational marginal pricing (LMP) data for all bus locations market and start_date to end_date and save it in folder. Parse the data using parser if required.

ElectricityMarketData.get_day_ahead_hourly_lmp_raw_data — Method

get_day_ahead_hourly_lmp_raw_data(market::PjmMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Day-Ahead Energy Market locational marginal pricing (LMP) data for all bus locations market and start_date to end_date and save it in folder.

ElectricityMarketData.get_day_ahead_lmp — Method

get_day_ahead_lmp(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Day-Ahead (DA) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData.get_day_ahead_lmp — Method

get_day_ahead_lmp(market::MisoMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Day-Ahead (DA) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData.get_day_ahead_lmp_raw_data — Method

get_day_ahead_lmp_raw_data(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Day-Ahead (DA) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder. Parse the data using parser if provided.

ElectricityMarketData.get_day_ahead_lmp_raw_data — Method

get_day_ahead_lmp_raw_data(market::MisoMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Day-Ahead (DA) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder. Parse the data using parser if provided.

ElectricityMarketData.get_group_df — Method

get_group_df(df::DataFrame, value_keys::Vector{String})::Dict{String, DataFrame}

Group the DataFrame by pnode_id and return a Dict of DataFrames for each value key.

Arguments

df::DataFrame: The DataFrame to be grouped.

value_keys::Vector{String}: The keys to be grouped into a Dict of DataFrames.

Returns

Dict{String, DataFrame}: A Dict of DataFrames for each value key.

ElectricityMarketData.get_meta_df — Method

get_meta_dict(df::DataFrame, meta_keys::Vector{String})::DataFrame

Get the meta DataFrame from the DataFrame.

Arguments

df::DataFrame: The DataFrame to get the meta DataFrame from.

meta_keys::Vector{String}: The keys to be included in the meta DataFrame.

Returns

DataFrame: The meta DataFrame.

ElectricityMarketData.get_real_time_lmp — Method

get_real_time_lmp(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Real-Time (RT) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData.get_real_time_lmp — Method

get_real_time_lmp(market::MisoMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Real-Time (RT) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData.get_real_time_lmp_raw_data — Method

get_real_time_lmp_raw_data(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Real-Time (RT) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder. Parse the data using parser if provided.

ElectricityMarketData.get_real_time_lmp_raw_data — Method

get_real_time_lmp_raw_data(market::MisoMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Real-Time (RT) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder. Parse the data using parser if provided.

ElectricityMarketData.get_str_dates — Method

get_str_dates(start_date::ZonedDateTime, end_date::ZonedDateTime, market::PjmMarket)::Tuple{String, String, String, String}

Get the start and end date strings in the format "m/d/yyyy" and the start and end hour and minute strings in the format "HH:MM".

ElectricityMarketData.get_timezone — Method

get_timezone(market::MisoMarket) :: TimeZone

Return the timezone of the given market.

ElectricityMarketData.get_timezone — Method

get_timezone(market::PjmMarket) :: TimeZone

Return the timezone of the given market.

ElectricityMarketData.get_timezone — Method

get_timezone(market::ElectricityMarket) :: TimeZone

Return the timezone of the given market.

ElectricityMarketData.get_url_day_ahead_hourly_lmp — Method

get_url_day_ahead_hourly_lmp(start_date::String, end_date::String)::AbstractString

Get the url for the day ahead hourly lmp data from the PJM API.

Arguments

start_date::String: The start date of the data to be retrieved.

end_date::String: The end date of the data to be retrieved.

starthourminute::String: The start hour and minute of the data to be retrieved.

endhourminute::String: The end hour and minute of the data to be retrieved.

Returns

AbstractString: The url for the day ahead hourly lmp data from the PJM API.

ElectricityMarketData.list_markets — Method

list_markets() :: Vector{ElectricityMarket}

Return a vector of all available electricity markets.

ElectricityMarketData.parse_df_format — Method

parsedfformat(df::DataFrame, valuekeys::Vector{String}, metakeys::Vector{String})::Tuple{Dict, DataFrame}

parse the DataFrame into a Dict of DataFrames for each value key and a DataFrame for the meta keys.

Arguments

df::DataFrame: The DataFrame to be parsed.

value_keys::Vector{String}: The keys to be parsed into a Dict of DataFrames.

meta_keys::Vector{String}: The keys to be parsed into a DataFrame.

Returns

Tuple{Dict, DataFrame}: A Tuple containing the Dict of DataFrames for each value key and a DataFrame for the meta keys.