ElectricityMarketData

Documentation for ElectricityMarketData.

ElectricityMarketData.ElectricityMarket

TimeZones.ZonedDateTime

ElectricityMarketData._async_download

ElectricityMarketData._async_download

ElectricityMarketData._async_get_raw_data

ElectricityMarketData._download

ElectricityMarketData.available_time_series

ElectricityMarketData.available_time_series

ElectricityMarketData.get_day_ahead_lmp

ElectricityMarketData.get_day_ahead_lmp_raw_data

ElectricityMarketData.get_real_time_lmp

ElectricityMarketData.get_real_time_lmp

ElectricityMarketData.get_real_time_lmp_raw_data

ElectricityMarketData.get_real_time_lmp_raw_data

ElectricityMarketData.get_timezone

ElectricityMarketData.get_timezone

ElectricityMarketData.list_markets

ElectricityMarketData.ElectricityMarket — Type

ElectricityMarket

Abstract type representing an electricity market.

TimeZones.ZonedDateTime — Method

ZonedDateTime(date::DateTime, timezone::TimeZone) :: ZonedDateTime

Return a ZonedDateTime object with the given date and timezone from the market.

ElectricityMarketData._async_download — Method

_async_download(url::AbstractString, filename::AbstractString)::Task

Downloads the 'url' and saves it to 'filename' asynchronously. Return the tasks.

ElectricityMarketData._async_download — Method

function _async_download(urls::Vector{T}, filenames::Vector{T}, url_zip::AbstractString, filename_zip::AbstractString)::Vector{Task} where T <: AbstractString

Downloads the 'urls' and saves it to 'filenames' asynchronously. If the file is not available, uses the zipped version. Return the tasks.

ElectricityMarketData._async_get_raw_data — Method

_async_get_raw_data(market::MisoMarket, url_folder::AbstractString, file_base::AbstractString, extension::AbstractString, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir())

Download the 'filebase' raw data for the given market in the given 'urlfolder' and start_date to end_date and save it in folder.

ElectricityMarketData._download — Method

_download(url::AbstractString, filename::AbstractString)::AbstractString

Downloads the return of url if the file filename does not exist, and saves it in filename.

ElectricityMarketData.available_time_series — Method

available_time_series(::MisoMarket) :: Vector{NamedTuple}

Return Vector of NamedTuple of available time series for the given market.

Ex:

[
    (name="RT-load", unit="MW", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_load_data, description="Real-time load data"),
    (name="RT-LMP", unit="MWh", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_lmp, description="Real-time Locational Marginal Price data"),
]

ElectricityMarketData.available_time_series — Method

available_time_series(market::ElectricityMarket) :: Vector{NamedTuple}

Return Vector of NamedTuple of available time series for the given market.

Ex:

[
    (name="RT-load", unit="MW", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_load_data, description="Real-time load data"),
    (name="RT-LMP", unit="MWh", resolution=Hour(1), first_date=DateTime("2021-01-01T00:00:00"), method=get_real_time_lmp, description="Real-time Locational Marginal Price data"),
]

ElectricityMarketData.get_day_ahead_lmp — Method

get_day_ahead_lmp(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Day-Ahead (DA) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData.get_day_ahead_lmp_raw_data — Method

get_day_ahead_lmp_raw_data(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Day-Ahead (DA) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder. Parse the data using parser if provided.

ElectricityMarketData.get_real_time_lmp — Method

get_real_time_lmp(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Real-Time (RT) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData.get_real_time_lmp — Method

get_real_time_lmp(market::MisoMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing) :: Tables.table

Return a table with Real-Time (RT) Locational Marginal Price (LMP) data for the given market and start_date to end_date. Parse the data using parser if provided. If the data is not available, download it and save it in folder.

ElectricityMarketData.get_real_time_lmp_raw_data — Method

get_real_time_lmp_raw_data(market::ElectricityMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Real-Time (RT) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder. Parse the data using parser if provided.

ElectricityMarketData.get_real_time_lmp_raw_data — Method

get_real_time_lmp_raw_data(market::MisoMarket, start_date::ZonedDateTime, end_date::ZonedDateTime; folder::AbstractString=tempdir(), parser::Function=(args...) -> nothing)

Download raw data for Real-Time (RT) Locational Marginal Price (LMP) for the given market and start_date to end_date and save it in folder. Parse the data using parser if provided.

ElectricityMarketData.get_timezone — Method

get_timezone(market::MisoMarket) :: TimeZone

Return the timezone of the given market.

ElectricityMarketData.get_timezone — Method

get_timezone(market::ElectricityMarket) :: TimeZone

Return the timezone of the given market.

ElectricityMarketData.list_markets — Method

list_markets() :: Vector{ElectricityMarket}

Return a vector of all available electricity markets.